requirements.txt
Django==4.0.1
django-environ==0.8.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
social-auth-app-django==5.0.0
requests==2.27.1
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
django-filter==21.1
drf-yasg==1.20.0
pytest-django==4.5.2
pydantic==1.9.2

